refactor(forms): Allow lazy loading of signal forms when using `provideSignalFormsConfig`
`provideSignalFormsConfig` is moved into its own chunk to prevent its usage from pulling the whole signal form code base.

fixes #67909
