+++ packages/forms/signals/src/directive/form_field.ts
@@ -36,7 +36,7 @@ import {type ValidationError} from '../api/rules';
 import type {Field, FieldState} from '../api/types';
 import {InteropNgControl} from '../controls/interop_ng_control';
 import {RuntimeErrorCode} from '../errors';
-import {SIGNAL_FORMS_CONFIG} from '../field/di';
+import {SIGNAL_FORMS_CONFIG} from '../config/di';
 import type {FieldNode} from '../field/node';
 import type {FormUiControl} from '../api/control';
 import {shallowArrayEquals} from '../util/array';
